Refactured names, found some spaces instead of tabs
