+++ Assets/Scripts/HighScoreEntry.cs
